microservice_aa/requirements.txt
aiormq==6.7.7
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
carrot-rpc==0.3.1
click==8.1.7
executing==1.2.0
fastapi==0.102.0
h11==0.14.0
httptools==0.6.0
idna==3.4
multidict==6.0.4
orjson==3.9.5
pamqp==3.2.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
simple-print==1.7.5
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
termcolor==1.1.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
